Verify the inclusion proof before forwarding msg
More work is required to tidy this up. At the moment the public key generated during my deployment is baked into the client. Depending on how the demo needs to be structured, we can either read the PK and replace it using the command provided, or we can fix a private/public key pair and create the tree with it in the first place.